+++ test/put_bit.c
@@ -22,7 +22,7 @@ int main(void)
   mkdir(filedir, 0777);
 
   for (i = 0; i < 8; ++i)
-    c[i] = i;
+    c[i] = (uint8_t)i;
 
   fd = open(format, O_CREAT | O_EXCL | O_WRONLY, 0666);
   write(fd, format_data, strlen(format_data));
